requirements.txt
streamlit 
langchain-community
langchain-groq
langchain-ollama
python-dotenv